[MISC] move common_tuple to correct place
