+++ include/seqan3/utility/views/pairwise_combine.hpp
@@ -15,9 +15,9 @@
 #include <cmath>
 #include <seqan3/std/ranges>
 
-#include <seqan3/core/tuple/common_tuple.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/utility/range/concept.hpp>
+#include <seqan3/utility/tuple/common_tuple.hpp>
 #include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
 
 namespace seqan3::detail
