+++ test/unit/alignment/matrix/detail/affine_cell_proxy_test.cpp
@@ -8,8 +8,8 @@
 #include "gtest/gtest.h"
 
 #include <seqan3/alignment/matrix/detail/affine_cell_proxy.hpp>
-#include <seqan3/core/tuple/common_tuple.hpp>
 #include <seqan3/test/expect_same_type.hpp>
+#include <seqan3/utility/tuple/common_tuple.hpp>
 
 //------------------------------------------------------------------------------
 // score cell proxy
